py-model/requirements.txt
fastapi
uvicorn[standard]
torch
torchvision
timm
ultralytics
transformers
pillow
numpy
python-multipart

